Add user check in server for deleteing Keeps

+++ Repositories/KeepsRepository.cs
@@ -59,10 +59,10 @@ namespace Keepr.Repositories
       _db.Execute(sql, update);
     }
 
-    internal void Delete(int id)
+    internal void Delete(int id, string userId)
     {
-      string sql = "DELETE FROM keeps WHERE id = @id";
-      _db.Execute(sql, new { id });
+      string sql = "DELETE FROM keeps WHERE (id = @id AND userId = @userId)";
+      _db.Execute(sql, new { id, userId });
     }
   }
 }
